requirements.txt
discord.py
PyNaCl
python-dotenv
openai==0.28.0
langdetect
Pillow